restruct code to reduce circle import.
