+++ torch/_inductor/cpp_builder.py
@@ -25,10 +25,9 @@ from torch._inductor.codecache import (
     _transform_cuda_paths,
     cpp_prefix_path,
     get_lock_dir,
-    invalid_vec_isa,
     LOCK_TIMEOUT,
-    VecISA,
 )
+from torch._inductor.cpu_vec_isa import invalid_vec_isa, VecISA
 from torch._inductor.runtime.runtime_utils import cache_dir
 
 if config.is_fbcode():
